+++ include/beast/http/message.hpp
@@ -322,7 +322,7 @@ struct header<false, Fields> : Fields
     */
     void
     reason(string_view s);
-   
+
 private:
 #if ! BEAST_DOXYGEN
     template<bool, class, class>
@@ -468,6 +468,20 @@ struct message : header<isRequest, Fields>
     message(std::piecewise_construct_t,
         std::tuple<BodyArgs...>&& body_args,
             std::tuple<HeaderArgs...>&& header_args);
+
+    /// Returns the header portion of the message
+    header_type const&
+    header_part() const
+    {
+        return *this;
+    }
+
+    /// Returns the header portion of the message
+    header_type&
+    header_part()
+    {
+        return *this;
+    }
 
     /// Returns `true` if "close" is specified in the Connection field.
     bool