+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Version 61:
 * Use generic_cateogry for errno
 * Reorganize SSL examples
 * Tidy up some integer conversion warnings
+* Add message::header_part()
 
 --------------------------------------------------------------------------------